+++ src/Joint_TEST.cc
@@ -29,6 +29,7 @@ TEST(DOMJoint, Construction)
   EXPECT_TRUE(joint.ChildLinkName().empty());
   EXPECT_EQ(ignition::math::Pose3d::Zero, joint.Pose());
   EXPECT_TRUE(joint.PoseFrame().empty());
+  EXPECT_EQ(nullptr, joint.Element());
 
   joint.SetPose({-1, -2, -3, IGN_PI, IGN_PI, 0});
   EXPECT_EQ(ignition::math::Pose3d(-1, -2, -3, IGN_PI, IGN_PI, 0),