+++ Changelog.md
@@ -9,6 +9,9 @@
 
 ### SDFormat 7.0.0 (xxxx-xx-xx)
 
+1. Preserve XML elements that are not part of the SDF specification.
+   * [Pull request 449](https://bitbucket.org/osrf/sdformat/pull-requests/449)
+
 1. Embed SDF specification files directly in libsdformat.so.
    * [Pull request 434](https://bitbucket.org/osrf/sdformat/pull-requests/434)
 
@@ -21,8 +24,10 @@
    * [Pull request 441](https://bitbucket.org/osrf/sdformat/pull-requests/441)
 
 1. SDF DOM: Additions to the document object model.
-   * [Pull request
-   442](https://bitbucket.org/osrf/sdformat/pull-requests/442)
+   * [Pull request 442](https://bitbucket.org/osrf/sdformat/pull-requests/442)
+
+1. SDF DOM: Add Element() accessor to Gui, JointAxis and World classes.
+   * [Pull request 450](https://bitbucket.org/osrf/sdformat/pull-requests/450)
 
 ## SDFormat 6.0